+++ bindings/python/CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Optional `max_parallel_api_requests` to `Client` constructor;
-- `Utils::{verify_transaction_syntactic(), block_bytes(), block_hash_without_nonce()}`;
+- `Utils::{verify_transaction_syntax(), block_bytes(), block_hash_without_nonce()}`;
 - `NodeInfoProtocol.belowMaxDepth`;
 - `{NodeInfoProtocol, RentStructure}::as_dict()`;